// src/kafka/topics/mt5/order.topics.enum.ts
export enum OrderTopics {
  orderOpenTicket = 'order-open-ticket',
  orderOpenTotal = 'order-open-total',
  orderOpenPaged = 'order-open-paged',
  orderOpenMultiple = 'order-open-multiple',
  updateOrder = 'update-order',
  deleteOrder = 'delete-order',
  cancelOrder = 'cancel-order',
  orderCloseTicket = 'order-close-ticket',
  orderCloseTotal = 'order-close-total',
  orderClosePaged = 'order-close-paged',
  orderClosedMultiple = 'order-closed-multiple',
  updateClosed = 'update-closed',
  deleteClosed = 'delete-closed',
  getBackupList = 'get-backup-list',
  getOrderFromBackup = 'get-order-from-backup',
  restoreOrderFromBackup = 'restore-order-from-backup',
  reopenOrder = 'reopen-order',
}